Update lyrics.lua
Ton of updates in this one.  Prepared Lyrics and Scene Loaded Lyrics never really played well in the sandbox.  Update has Scene Loaded lyrics always in the first prepared lyrics position, and this first position song NOT saved to the file if loaded by a scene.  Having the monitor window exposed a lot of weird index positions that worked but were not really correct, so these also got cleaned up so Song x of n and Page x of n all show as expected in the monitor.  The monitor window acted a bit like a missing "debugger" window for Lua and that really helped.