Updates Fade and other stuff
This is another branch I was working on to enable working with Studio Mode and duplicate sources.  Using Dup Sources allows the lyrics to change in preview but not in program.   A checkbox in properties allows for the automatic transition from preview to program after a lyrics change if desired. This auto transition is mutually exclusive with text fade option just to keep people out of trouble.  Dup Sources also allows for OBS transitions to apply to text sources or entire scene.   It was almost working and did not have the fade issues, so rather than back up, I went forward and finished this branch.  I hope.