OBS workaround for Duplicate Sources
It seems OBS does not offer a way to look inside a set of duplicate sources created to separate preview from program in Studio Mode.  As a workaround, I implemented a FirstTransition toggle that gets reset on preparing a new song and set once a manual Next Lyric function is performed.  That first NEXT only does a transition.  It is only to keep the AUTO transition from working until signaled to do so.  Pretty advanced for most people , but due to some camera option scripting I needed it to work with duplicate sources.