Update lyrics.lua
removed the Unprepare button and added Edit List Button which allows for deleting and reordering.  You could add a prepared song here but it would need to match the name exactly.  I think I can add a test that every prepared song in the list is actually in the directory.   Comments welcome on this one.