Update lyrics.lua
WooHoo   I found a way to use the save iteration for renaming!  Much more efficient than the old way.  I also removes the get_ID call which some documents have said to NOT USE because of instability it causes in OBS.  Create adds a new isLLC true bool data item to settings in each source that can be checked to see when a source is ours.