Update lyrics.lua
Moved some methods around to follow them better.  Not that many changes.  Added two more status to text.   TEXT_HIDE and TEXT_SHOW

Only using TEXT_HIDE for now. but both instantly change the text visibility and ignore the fade if set.  All the other end states honor fades.

When transitioning via a source that has loaded the text in preview, it needs to instantly turn off the text before the transition so it can play nice and fade in again.

Okay back to documenting.