Some fixes for fading non-100% opacities.
It is possible to get a text object "lost" by having it fade to 0% with HIDE then closing the script without putting it back, when not forcing 0-100% fades.  This is because the script needs to try to "read" the current opacity level and then re-use that as the max.  So a new button allows manually setting these maximums when all sources are as they should be.  Doing this automatically is nearly impossible.  Probably setting source item callbacks and trying to management them. I have scripts that do this but they are a pain to keep up with.  Also required is the option to NOT fade backgrounds if you want to keep them transparent for text items for example, or just want them left alone.  I employed progressive disclosure whenever possible to try to limit complexity of the UI until required.